12/12 Cositas #1

+++ servicioSalud/src/main/java/com/app/servicioSalud/entidades/Calificacion.java
@@ -1,3 +1,7 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
 package com.app.servicioSalud.entidades;
 
 import javax.persistence.Entity;
@@ -9,20 +13,29 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import lombok.Data;
 
+/**
+ *
+ * @author EduRiu
+ */
 @Data
 @Entity
 public class Calificacion {
-
+    
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
+    private Integer id;
+    
     @ManyToOne(fetch = FetchType.EAGER)
-    @JoinColumn(name = "profesional_id")
+    @JoinColumn (name = "profesional_id")
     private Profesional profesional;
-
+    
     private Double puntualidad;
     private Double mobiliario;
     private Double atencion;
-    private String comentario;
+    private String comentario; 
+          
+    
+    
+     
+    
 }
